requirements.txt
    # requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
nats-py==2.5.0
requests==2.31.0
pyyaml==6.0.1
pydantic==2.5.0
pydantic[email]==2.5.0  # This includes email-validator
python-multipart==0.0.6
httpx==0.25.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi 
opentelemetry-instrumentation-uvicorn 
opentelemetry-instrumentation-httpx 
opentelemetry-instrumentation-redis 
opentelemetry-instrumentation-sqlalchemy 
opentelemetry-instrumentation-logging
opentelemetry-bootstrap -a install
fastapi==0.104.1
uvicorn[standard]==0.24.0
nats-py==2.5.0
pydantic==2.4.2
python-multipart==0.0.6
requests==2.31.0
black==23.11.0 